// src/app/models/book.ts
export class Book {

    id: number;

    title: string;

    author: string;

    publisher: string;

    publicationDate: string;

    language: string;

    category: string;

    pages: number;

    format: string;

    isbn: string;

    shippingWeight: number;

    listPrice: number;

    ourPrice: number;

    active: boolean;

    stockNumber: number;

    photo: string;

    description: string;
}